+++ packages/pg/lib/client.js
@@ -1,5 +1,3 @@
-'use strict'
-
 const EventEmitter = require('events').EventEmitter
 const utils = require('./utils')
 const nodeUtils = require('util')
@@ -14,23 +12,28 @@ const crypto = require('./crypto/utils')
 
 const activeQueryDeprecationNotice = nodeUtils.deprecate(
   () => {},
-  'Client.activeQuery is deprecated and will be removed in a future version.'
+  'Client.activeQuery is deprecated and will be removed in pg@9.0'
 )
 
 const queryQueueDeprecationNotice = nodeUtils.deprecate(
   () => {},
-  'Client.queryQueue is deprecated and will be removed in a future version.'
+  'Client.queryQueue is deprecated and will be removed in pg@9.0.'
 )
 
 const pgPassDeprecationNotice = nodeUtils.deprecate(
   () => {},
-  'pgpass support is deprecated and will be removed in a future version. ' +
-    'You can provide an async function as the password property to the Client/Pool constructor that returns a password instead. Within this funciton you can call the pgpass module in your own code.'
+  'pgpass support is deprecated and will be removed in pg@9.0. ' +
+    'You can provide an async function as the password property to the Client/Pool constructor that returns a password instead. Within this function you can call the pgpass module in your own code.'
 )
 
 const byoPromiseDeprecationNotice = nodeUtils.deprecate(
   () => {},
-  'Passing a custom Promise implementation to the Client/Pool constructor is deprecated and will be removed in a future version.'
+  'Passing a custom Promise implementation to the Client/Pool constructor is deprecated and will be removed in pg@9.0.'
+)
+
+const queryQueueLengthDeprecationNotice = nodeUtils.deprecate(
+  () => {},
+  'Calling client.query() when the client is already executing a query is deprecated and will be removed in pg@9.0. Use async/await or an external async flow control mechanism instead.'
 )
 
 class Client extends EventEmitter {
@@ -286,7 +289,7 @@ class Client extends EventEmitter {
     if (typeof this.password === 'function') {
       this._Promise
         .resolve()
-        .then(() => this.password())
+        .then(() => this.password(this.connectionParameters))
         .then((pass) => {
           if (pass !== undefined) {
             if (typeof pass !== 'string') {
@@ -882,8 +885,12 @@ class Client extends EventEmitter {
       }
       readTimeout = config.query_timeout || this.connectionParameters.query_timeout
       result = query = config
-      if (typeof values === 'function') {
-        query.callback = query.callback || values
+      if (!query.callback) {
+        if (typeof values === 'function') {
+          query.callback = values
+        } else if (callback) {
+          query.callback = callback
+        }
       }
     } else {
       readTimeout = config.query_timeout || this.connectionParameters.query_timeout
@@ -980,6 +987,9 @@ class Client extends EventEmitter {
       return result
     }
 
+    if (this._queryQueue.length > 0) {
+      queryQueueLengthDeprecationNotice()
+    }
     this._queryQueue.push(query)
     this._pulseQueryQueue()