+++ CHANGELOG.md
@@ -4,6 +4,11 @@ For richer information consult the commit log on github with referenced pull req
 
 We do not include break-fix version release in this file.
 
+## pg@8.19.0
+
+- [Deprecate interal query queue](https://github.com/brianc/node-postgres/pull/3603).
+- Pass connection parameters [to password callback](https://github.com/brianc/node-postgres/pull/3602).
+
 ## pg@8.18.0
 
 - [Return the client instance](https://github.com/brianc/node-postgres/pull/3564) as the result of calling `connect` (previously it was `void`).